Add files via upload
stack overflow survey project
we are going to work on Visualizations. Our task is to analyze the data, clean it, modify it as and when required and represent it in the form of graphs, charts, plots, which are more appealing to users.
For this project we will be exploring data from Developers’ Survey done by stackoverflow.com every year. This data is huge consisting of around 1 lac developers’ world-wide and on multiple factors.